refactor(chat): relocate WindowFrameContext, rename ResourceListUiService, fix stale comments
Addresses @0xfullex's review (#16272):
- I1: move WindowFrameContext from the top-level context/ bucket (flagged for
  dissolution) into its owning domain components/chat/shell/; repoint imports.
- I2: rename the state-owning class ResourceListUiStore -> ResourceListUiService
  (naming-conventions §5.2: Service for stateful classes, not a Store suffix);
  the useResourceListUiStore hook / context / store var keep their names.
- Nit: drop dangling comment references (non-existent buildFileTreeNodes /
  flat-paths era in ArtifactPane; "backward compatibility" on a new wrapper).
