+++ src/renderer/components/chat/panes/Shell/Shell.tsx
@@ -1,9 +1,9 @@
 import { HorizontalScrollContainer, Tabs, TabsContent, TabsList, TabsTrigger, Tooltip } from '@cherrystudio/ui'
+import { useWindowFrame } from '@renderer/components/chat/shell/WindowFrameContext'
 import { CommandTooltip } from '@renderer/components/command'
 import { RightSidebarCollapseIcon, RightSidebarExpandIcon } from '@renderer/components/Icons'
 import NavbarIcon from '@renderer/components/NavbarIcon'
 import { isMac } from '@renderer/config/constant'
-import { useWindowFrame } from '@renderer/context/WindowFrameContext'
 import { useCommandHandler } from '@renderer/hooks/command'
 import { cn } from '@renderer/utils'
 import type { CommandId } from '@shared/utils/command/definitions'
