+++ src/renderer/components/chat/shell/WindowFrameContext.tsx
